move entity struct to entity folder
