+++ login/signout.go
@@ -7,6 +7,7 @@ import (
 	"github.com/linkernetworks/logger"
 	response "github.com/linkernetworks/net/http"
 	"github.com/linkernetworks/session"
+	"github.com/linkernetworks/webservice/login/entity"
 )
 
 func (s *LoginService) signOut(req *restful.Request, resp *restful.Response) {
@@ -24,7 +25,7 @@ func (s *LoginService) signOut(req *restful.Request, resp *restful.Response) {
 		return
 	}
 
-	resp.WriteHeaderAndEntity(http.StatusOK, SignInResponse{
+	resp.WriteHeaderAndEntity(http.StatusOK, entity.SignInResponse{
 		Error:     false,
 		Message:   "Logout success",
 		SignInUrl: "/signin",
