+++ login/signup.go
@@ -9,6 +9,7 @@ import (
 	oauth "github.com/linkernetworks/oauth/entity"
 	"github.com/linkernetworks/oauth/util"
 	"github.com/linkernetworks/validator"
+	"github.com/linkernetworks/webservice/login/entity"
 	"github.com/linkernetworks/webservice/pwdutil"
 	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
@@ -63,7 +64,7 @@ func (s *LoginService) signUp(req *restful.Request, resp *restful.Response) {
 		validations["password"] = passworkValidate
 	}
 	if validations.HasError() {
-		resp.WriteHeaderAndEntity(http.StatusUnprocessableEntity, ActionResponse{
+		resp.WriteHeaderAndEntity(http.StatusUnprocessableEntity, entity.ActionResponse{
 			Error:       true,
 			Validations: validations,
 			Message:     "Input data is not valid",
@@ -88,7 +89,7 @@ func (s *LoginService) signUp(req *restful.Request, resp *restful.Response) {
 		response.InternalServerError(req.Request, resp.ResponseWriter, err)
 		return
 	}
-	resp.WriteEntity(ActionResponse{
+	resp.WriteEntity(entity.ActionResponse{
 		Error:   false,
 		Message: "Sign up success",
 	})
